Simplify eager loops so they are properly converted to poplar repeat
Summary:
New keras API used a different way of looping over the training steps causing it not to be properly converted to a `poplar::repeat`. Instead we ended up with a `repeatWhile` instead of just a `repeat` causing an increase in memory.

TF2.5 Only

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, babakk, alfiee

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, vladimirm, alfiee

Maniphest Tasks: T55704

Differential Revision: https://phabricator.sourcevertex.net/D60567

+++ tensorflow/python/ipu/keras/extensions/keras_extension_base.py
@@ -432,7 +432,7 @@ class KerasExtensionBase(base_layer.KerasExtension):
 
     @def_function.function(experimental_compile=True)
     def train_function(steps_per_execution, iterator, outfeed):
-      for _ in math_ops.range(np.int64(steps_per_execution)):
+      for _ in math_ops.range(steps_per_execution):
         outfeed.enqueue(train_step(next(iterator)))
 
     return train_function
